Rename to Betwixt
Bang a la clang was a horrible name that attempted to satisfy Tim's
desire for naming continuity with Flashbang with my desire to not have
flash in the name. clangbang would've been a better middle ground in
that direction, if not for its also sounding like it's a library for
producing robot porn.

It's now called Betwixt because the library is about being between your
app and the framework. Tim still wanted the name to be Flashbang, but
I won the battle of bikeshedding stamina. He maintains that time will
prove me a fool.