Add named nodes to all contexts
Nodes can be given a name when added to a context, or they can replace an existing item with that name in the context. The names only need to be unique in that context.

Mode-wide names are now called keys. If an object conforms to BTKeyed, it's added to the keyed lookup on the mode it's attached to.